refactor: [170] removed utils folder, extracted files in the root of nn
