+++ fenn/__init__.py
@@ -1,8 +1,7 @@
 from fenn.agents import Flow, Node, RAGNode
 from fenn.agents.llm import LLMClient
 from fenn.app import Fenn
-from fenn.nn import Checkpoint, ClassificationTrainer, RegressionTrainer
-from fenn.nn.trainers import LoRATrainer
+from fenn.nn import Checkpoint, ClassificationTrainer, LoRATrainer, RegressionTrainer
 from fenn.reproducibility import set_seed
 
 __all__ = [
