+++ fenn/nn/__init__.py
@@ -1,4 +1,14 @@
-from .trainers import ClassificationTrainer, RegressionTrainer, Trainer
-from .utils import Checkpoint
+from .checkpoint import Checkpoint
+from .model_pretty_printer import ModelPrettyPrinter
+from .state import TrainingState
+from .trainers import ClassificationTrainer, LoRATrainer, RegressionTrainer, Trainer
 
-__all__ = ["Trainer", "ClassificationTrainer", "RegressionTrainer", "Checkpoint"]
+__all__ = [
+    "ClassificationTrainer",
+    "LoRATrainer",
+    "RegressionTrainer",
+    "Trainer",
+    "Checkpoint",
+    "TrainingState",
+    "ModelPrettyPrinter",
+]
