+++ fenn/nn/state.py
