+++ fenn/nn/trainers/regression_trainer.py
@@ -15,7 +15,7 @@ from sklearn.metrics import r2_score
 from torch.utils.data import DataLoader
 
 from fenn.logging import logger
-from fenn.nn.utils import Checkpoint
+from fenn.nn.checkpoint import Checkpoint
 
 from .trainer import Trainer
