--- fenn/nn/utils/__init__.py
@@ -1,5 +1,0 @@
-from .checkpoint import Checkpoint
-from .model_pretty_printer import ModelPrettyPrinter
-from .state import TrainingState
-
-__all__ = ["Checkpoint", "ModelPrettyPrinter", "TrainingState"]
